fix: Add missing `BUILD`
