+++ src/ai/backend/common/data/session/types.py
@@ -1,0 +1,10 @@
+import enum
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    pass
+
+
+class CustomizedImageVisibilityScope(str, enum.Enum):
+    USER = "user"
+    PROJECT = "project"
