+++ src/ai/backend/common/data/vfolder/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="src")
